pyproject.toml
[tool.poetry]
name = "complaints"
version = "0.1.0"
description = ""
authors = ["Mitch Murphy <mitch.murphy@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.78.0"
uvicorn = "^0.17.6"
asyncpg = "^0.25.0"
psycopg2-binary = "^2.9.3"
psycopg2 = "^2.9.3"
SQLAlchemy = "^1.4.37"
alembic = "^1.8.0"
databases = "^0.6.0"
python-dotenv = "^0.20.0"
PyJWT = "^2.4.0"
passlib = "^1.7.4"
bcrypt = "^3.2.2"
email-validator = "^1.2.1"
asyncclick = "^8.1.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
